refactor: 예제 서버 mock폴더 최상단으로 이동
예제 API 서버로 활용 가능하도록 수정할 예정
