+++ mock/server/middleware/adapter.go
